Maintain value-less attributes handled by JSX
Attributes with optional values, like `disabled` for example, are
assigned an empty string as their value making them evaluate as falsy.
In HTML, if the value is omitted the attribute defaults to truthy and so
should be the case in the transform.

Currently:

```
> converter.convert("<input disabled>")
"<input disabled="" />
"
```

What should happen:

```
> converter.convert("<input disabled>")
"<input disabled />
"
```

JSX properly converts attributes with no values to `true`.

Fixes #12.